Dimitrios Part Completed, ready for testing
All of the featurs that I mentioned have been implemented, now only the seekers can disband the group and delete it from the database after the 20 minute mark passes. I conducted tests to make sure that it works, and that the button only appears to the seeker. The hiders get appropriate messages as well. The timers work for each group as long as the time is running. It works perfect even if you leave, rejoin, refresh etc.